fix: allow decimal CT calibration values (ICal)
parseInt/Math.round was truncating decimal ICal values (e.g. 101.5 → 102)
when reading from device or selecting from UI. Replaced with parseFloat
throughout, and added step="0.1" to the custom CT number input.

Fixes #28

+++ custom_components/emontx_config/frontend/components/modals.js
@@ -201,7 +201,7 @@ Vue.component('bulk-ct-modal', {
                 this.ctTypeCustom = true;
             } else {
                 this.ctTypeCustom = false;
-                this.ctTypeValue = parseInt(value);
+                this.ctTypeValue = parseFloat(value);
             }
         },
         apply() {
